LayoutTests/accessibility/loading-iframe-updates-axtree-expected.txt

Before


After

End of test

This tests that if an iframe loads new content after its accessibility object has already been accessed, the iframe accessibility object's descendants are the new scroll area and web area, not the old deleted ones.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.isEqual(newIframe) is true
PASS scrollarea.isEqual(newScrollarea) is false
PASS subwebarea.isEqual(newSubwebarea) is false
PASS newSubwebarea.childrenCount > 0 is true

TEST COMPLETE
